Restructure k8s e2e test code
